+++ munetic_admin/src/pages/UserListPage.tsx
@@ -1,14 +1,13 @@
 import Button from '@mui/material/Button';
 import userDummy from '../dummy/userDummy';
+import List from '../components/List';
 import ListContainer from '../components/ListContainer';
-import ListUserCard from '../components/ListUserCard';
+import ListCard from '../components/ListCard';
 import ListUserColumn from '../components/ListUserColumn';
 import CustomPagination from '../components/CustomPagination';
 
 export default function UserListPage() {
-  const userList: JSX.Element[] = userDummy.map(user => (
-    <ListUserCard {...user} />
-  ));
+  const userList: JSX.Element[] = userDummy.map(user => <ListCard {...user} />);
 
   return (
     <div>
@@ -18,10 +17,7 @@ export default function UserListPage() {
       <Button sx={{ fontSize: 12 }} variant="outlined">
         export
       </Button>
-      <ListContainer>
-        <ListUserColumn />
-        {userList}
-      </ListContainer>
+      <List />
       <CustomPagination />
     </div>
   );
